chore: rename test paths
